+++ Driver/io/can/can.hpp
@@ -19,7 +19,7 @@ public:
     );
     ~Can();
 
-    bool sendFrame(const std::vector<uint8_t>& data, size_t id);
+    bool sendFrame(const std::vector<uint8_t>& data, unsigned int id);
     // TODO: 接受指定 canID 数据帧
     bool receiveFrame(std::vector<uint8_t>& data, std::shared_ptr<size_t> id = nullptr);
     bool isOpen() const;
